Update README.md
Updated instructions as they are not very clear on how/from where one should update the start gcode to allow trigger messages for postprocessing to be updated. In addition, the same effect could be had by replacing the trigger message with a similar trigger message present in the non-post-processing repository without adverse affects.